STM32: add basic kdebug support.
git-svn-id: https://src.develer.com/svnoss/bertos/trunk@3661 38d2e660-2303-0410-9eaa-f027e97ec537

+++ bertos/cpu/cortex-m3/io/stm32_memmap.h
@@ -50,7 +50,7 @@
 #define OB_BASE               (0x1FFFF800)
 
 /* Peripheral memory map */
-#define APB1PERIPH_BASE       ((uint32_t)PERIPH_BASE)
+#define APB1PERIPH_BASE       (PERIPH_BASE)
 #define APB2PERIPH_BASE       (PERIPH_BASE + 0x10000)
 #define AHBPERIPH_BASE        (PERIPH_BASE + 0x20000)
